perf: optimize PDF rendering performance and memory management
Refactors the rendering pipeline and layout logic to improve responsiveness during zoom and scroll operations. Key optimizations include a more aggressive tile caching strategy and better resource reclamation.

Core changes:
- **Rendering Optimization**: Introduced a dedicated thread pool for CPU-intensive rasterization and implemented dynamic parallel renderer scaling based on available processors.
- **Enhanced Tiling**: Replaced the `renderedTiles` map with a `tileCache` (LRU) and introduced `tileRevision` to trigger redraws without full recompositions. Tiles now use stepped zoom levels (sqrt(2)) to improve cache hits and visual stability.
- **Memory Management**: Improved `BitmapPool` integration with a delayed release mechanism to ensure bitmaps are not reclaimed while still being drawn by Compose.
- **Layout Improvements**: Optimized `visiblePageIndices` using binary search and moved coordinate calculations to the Draw phase where possible to minimize layout overhead.
- **Page Rendering**: Removed Night Mode filtering from the low-level renderer; it is now handled via `ColorFilter` at the UI layer for better performance.
- **Bug Fixes**: Refined gesture handling and pan clamping to prevent jitter during pinch-zoom and flings.

+++ composepdf/src/main/kotlin/com/composepdf/gesture/GestureModifiers.kt
@@ -116,8 +116,6 @@ internal fun rememberGestureState(): GestureState {
     return remember { GestureState(scope) }
 }
 
-// ── pdfGestures modifier ──────────────────────────────────────────────────────
-
 /**
  * Unified PDF gesture [Modifier] handling pinch-zoom, pan, fling, and double-tap.
  *
@@ -135,10 +133,10 @@ internal fun rememberGestureState(): GestureState {
  * ## Touch-slop handling
  *
  * Events are not forwarded to the controller until the accumulated pan (in pixels)
- * or zoom (converted to pixels via viewport width) exceeds [ViewConfiguration.touchSlop].
+ * or zoom (converted to pixels via viewport width) exceeds [androidx.compose.ui.platform.ViewConfiguration.touchSlop].
  * This prevents micro-jitter from triggering scrolls during taps.
  *
- * ## Why [awaitEachGesture] instead of [detectTransformGestures]?
+ * ## Why [awaitEachGesture] instead of [androidx.compose.foundation.gestures.detectTransformGestures]?
  *
  * `detectTransformGestures` does not expose the raw pointer stream needed for
  * double-tap detection and velocity tracking. [awaitEachGesture] gives us full
@@ -256,14 +254,12 @@ fun Modifier.pdfGestures(
                     // Zoom end - check if we need to snap or just stop
                     controller.onGestureEnd()
                 } else {
-                    // Fling?
                     val velocity = gs.velocityTracker.calculateVelocity()
-                    val maxVelocity = viewConfiguration.maximumFlingVelocity.toFloat()
-                    val minVelocity = viewConfiguration.minimumFlingVelocity.toFloat() // 50dp/s approx
+                    val maxVelocity = viewConfiguration.maximumFlingVelocity
+                    val minVelocity = viewConfiguration.minimumFlingVelocity
 
                     val velocityX = velocity.x.coerceIn(-maxVelocity, maxVelocity)
                     val velocityY = velocity.y.coerceIn(-maxVelocity, maxVelocity)
-                    // Velocity no tiene .getDistance(), calculamos la magnitud manualmente
                     val speed = kotlin.math.sqrt(velocity.x * velocity.x + velocity.y * velocity.y)
 
                     if (speed > minVelocity) {
